Christen the new build server
Due to the new server using a brand new image, some of the dependency
assumptions changed. `bzip2-devel` is no longer in the base imagery or
`system.devel` component, so we must explicitly ask for it nowadays.

This build feels really expensive when you consider the VAT cost of the
new server today, and I mean, it's only php. :P

Signed-off-by: Ikey Doherty <ikey@solus-project.com>